Added Login Signup Features

+++ frontend/src/components/WorkoutDetails.js
@@ -1,5 +1,6 @@
 import { useWorkoutsContext } from "../hooks/useWorkoutsContext";
 import { Link } from "react-router-dom";
+import { useAuthContext } from "../hooks/useAuthContext";
 
 //date fns
 import { formatDistanceToNow } from 'date-fns';
@@ -7,10 +8,19 @@ import { formatDistanceToNow } from 'date-fns';
 const WorkoutDetails = ({ workout }) => {
 
     const { dispatch } = useWorkoutsContext();
+    const { user } = useAuthContext();
 
     const handleClick = async () => {
+        if (!user)
+        {
+            return;
+        }
+
         const response = await fetch('/api/workouts/' + workout._id, {
-            method: 'DELETE'
+            method: 'DELETE',
+            headers: {
+                'Authorization': `Bearer ${user.token}`
+            }
         });
 
         const data = await response.json();
